requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
supabase==2.7.4
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.9
google-generativeai>=0.8.0
pydantic>=2.0.0
